started using vue

+++ app/Http/Controllers/TodosController.php
@@ -15,7 +15,7 @@ class TodosController extends Controller
     public function index()
     {
         //
-        return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
     }
 
     /**
@@ -42,7 +42,11 @@ class TodosController extends Controller
             'title' => 'required|string',
             'completed' => 'required|boolean'
         ]);
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed
+        ]);
         return response($todo, 201);
     }
 
@@ -57,6 +61,9 @@ class TodosController extends Controller
     public function update(Request $request, Todo $todo)
     {
         //
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
         $data = $request->validate([
             'title' => 'required|string',
             'completed' => 'required|boolean'
@@ -71,7 +78,7 @@ class TodosController extends Controller
             'completed' => 'required|boolean'
         ]);
         // update all!
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
         return response('Updated', 200);
     }
 
@@ -84,7 +91,9 @@ class TodosController extends Controller
     public function destroy(Todo $todo)
     {
         //
-
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
         $todo->delete();
         return response('Deleted todo item', 200);
     }
@@ -92,6 +101,17 @@ class TodosController extends Controller
     public function destroyCompleted(Request $request)
     {
         //
+
+        $todosToDelete = $request->todos;
+        $userTodoIds = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        });
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
         $request->validate([
             'todos' => 'required|array'
         ]);
